Make gRingCount a u16

+++ src/game/interactables_2/notification_ring_bonus.c
@@ -102,7 +102,7 @@ void sub_80803FC(Sprite_IaUnknown *sprite)
                 gRingCount = ringBonus;
 
                 if ((gCurrentLevel != LEVEL_INDEX(ZONE_FINAL, ACT_TRUE_AREA_53))
-                    && (Div((u16)gRingCount, 100) != Div(prevRingCount, 100))
+                    && (Div(gRingCount, 100) != Div(prevRingCount, 100))
                     && (gGameMode == GAME_MODE_SINGLE_PLAYER)) {
                     u32 lives = gNumLives + 1;
                     if (lives > 255)
@@ -114,7 +114,7 @@ void sub_80803FC(Sprite_IaUnknown *sprite)
                 }
 
                 if (gGameMode == GAME_MODE_MULTI_PLAYER_COLLECT_RINGS) {
-                    if ((u16)gRingCount > 255) {
+                    if (gRingCount > 255) {
                         gRingCount = 255;
                     }
                 }
